+++ R/check_git_cloning.R
@@ -4,7 +4,7 @@
 #' @export
 check_git_cloning = R6::R6Class(
   "check_git_cloning",
-  inherit = uatBase::base_check,
+  inherit = audit.base::base_check,
   public = list(
     #' @description Checks deployment of a Plumber API
     #' @param debug_level See check() for details