+++ R/create_config.R
@@ -16,5 +16,5 @@
 #' * force: overwrite existing file
 #' * error: if a config file exists, raise an error
 #' @export
-create_config = uatBase::create_config(file = "config-uat-psw.yml",
+create_config = audit.base::create_config(file = "config-uat-psw.yml",
                                        pkg_name = "jrHealthCheckWorkbench")